lessen perf impact from firing guns
We do this by 1) putting ephemeral mobjs into a separate array (so svelte
doesn't have to re-evaluate the whole list on add/remove -- svelte 5
may help here) and 2) adding markup into the canvas to reduce reflow.

I'm not sure why reflow happens... svelte adds comments to the dom after
components but... reflow from comments?

On a large map, we used to drop 50% (eg. from 40fps to 20fps) but now the impact
is more like 20-25% (so 40fps to 28-32fps). I'd really like the impact to be closer
to 0 but don't have any more ideas at the moment