play music (sort-of)
It's almost enough to play E1 music but there is a lot to be desired. I'm
still a little confused at the midi instrument landscape in JS and the
solution I have here will download a ton of samples. Maybe it would be
better to download synth instruments instead? I'm still learning.

Even with this library, I'm not sure it's fast enough to keep up. The music
can sometimes pause or slow down a little in certain tracks (especially in firefox).