remove kaitai struct
It wasn't that hard to replace really and we've saved a bunch of code and
knocked off one dependency so it feels like a win all around.

I want to say this increased perf or cut down on memory usage but I'm not
confident in measuring that. Maybe someday I'll figure it out.