use events for sprite frame changes
For cosmogenesis map05, I go from 17-20fps to 50-60fps but I think the
event library is causing some slowdown. I'll see if there is a better on
or simply write my own. According to the profiler, we've cut time to update
sprites from 32% to 20% so the fps boost is surprising. I'm guessing
there is less memory used and that cost is less visible. Not sure